requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
build==0.10.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docutils==0.20.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
multidict==6.0.4
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pip-review==1.3.0
pip-tools==6.13.0
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
websockets==11.0.3
yarl==1.9.2
